+++ pennylane_lightning/lightning_tensor/_measurements.py
@@ -84,7 +84,9 @@ class LightningTensorMeasurements:
         return self._measurement_lightning.expval(ob_serialized)
 
     def var(self, measurementprocess: MeasurementProcess):
-        """Variance of the supplied observable contained in the MeasurementProcess.
+        """Variance of the supplied observable contained in the MeasurementProcess. Note that the variance is
+        calculated as <obs**2> - <obs>**2. The current implementation only supports single-wire observables.
+        2 and 2+-wires observables, projector and sparse-hamiltonian are not supported.
 
         Args:
             measurementprocess (StateMeasurement): measurement to apply to the state
@@ -92,7 +94,6 @@ class LightningTensorMeasurements:
         Returns:
             Variance of the observable
         """
-
         if isinstance(measurementprocess.obs, qml.SparseHamiltonian):
             raise NotImplementedError("Sparse Hamiltonian Observables are not supported.")